% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{all2int}
\alias{all2int}
\title{convert all discretized vectors to integer}
\usage{
all2int(x)
}
\arguments{
\item{x}{A discretized vector.}
}
\value{
An integer vector.
}
\description{
convert all discretized vectors to integer
}
\examples{
all2int(factor(letters[1:3],levels = c('b','a','c')))
all2int(letters[1:3])

}
