% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_spade.R
\name{F_informationloss}
\alias{F_informationloss}
\title{measure information loss by information entropy}
\usage{
F_informationloss(xvar, xdisc)
}
\arguments{
\item{xvar}{The original un-discretized vector.}

\item{xdisc}{The discretized vector.}
}
\value{
A value of information loss as measured by information entropy.
}
\description{
Function for measure information loss by information entropy.
}
\details{
The information loss measured by information entropy formula is
\eqn{F = -\sum\limits_{i=1}^N p_{(i)}\log_2 p_{(i)} - \left(-\sum\limits_{h=1}^L p_{(h)}\log_2 p_{(h)}\right)}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
