% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{gd_bestunidisc}
\alias{gd_bestunidisc}
\title{Best univariate discretize parameters based on geodetector q-statistic}
\usage{
gd_bestunidisc(
  formula,
  data,
  discnum,
  discmethod = NULL,
  cores = 1,
  return_disc = TRUE,
  ...
)
}
\arguments{
\item{formula}{A formula of spatial stratified heterogeneity test.}

\item{data}{A data.frame or tibble of observation data.}

\item{discnum}{A vector of number of classes for discretization.}

\item{discmethod}{(optional) A vector of methods for discretization,default is used
\code{c("sd","equal","pretty","quantile","fisher","headtails","maximum","box")}in \code{spEcula}.}

\item{cores}{positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}

\item{return_disc}{(optional) Whether or not return discretized result used the optimal parameter.
Default is \code{TRUE}.}

\item{...}{(optional) Other arguments passed to \code{st_unidisc()}.}
}
\value{
A list with the optimal parameter in the provided parameter combination with \code{k},
\code{method} and \code{disc}(when \code{return_disc} is \code{TRUE}).
}
\description{
Function for determining the best univariate discretize parameters based on geodetector q-statistic.
}
\examples{
\dontrun{
library(terra)
library(tidyverse)
fvcpath = system.file("extdata", "FVC.zip",package = 'spEcula')
fvc = terra::rast(paste0("/vsizip/",fvcpath))
fvc = as_tibble(terra::as.data.frame(fvc,na.rm = T))
g = gd_bestunidisc(fvc ~ .,data = select(fvc,-lulc),discnum = 2:15,cores = 6)
new.fvc = g$disv
new.fvc = bind_cols(select(fvc,fvc,lulc),new.fvc)
ssh.test(fvc ~ .,data = new.fvc,type = 'factor')
ssh.test(fvc ~ .,data = new.fvc,type = 'interaction')
}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
