% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gozh.R
\name{gd_rpart}
\alias{gd_rpart}
\title{q-statistics of geographical detector based on recursive partitioning}
\usage{
gd_rpart(formula, data, ...)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.frame or tibble of observation data.}

\item{...}{(optional) Other arguments passed to \code{rpart::rpart()}.}
}
\value{
A tibble contains the Q-statistic and the p-value.
}
\description{
q-statistics of geographical detector based on recursive partitioning
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
