% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_sesu.R
\name{gd_sesu}
\alias{gd_sesu}
\title{comparison of size effects of spatial units based on optimal parameters geographical detector(OPGD) model.}
\usage{
gd_sesu(
  formula,
  datalist,
  su,
  discvar,
  discnum = NULL,
  discmethod = NULL,
  cores = 1,
  ...
)
}
\arguments{
\item{formula}{A formula.}

\item{datalist}{A list of \code{data.frame} or \code{tibble}.}

\item{su}{A vector of sizes of spatial units.}

\item{discvar}{Name of continuous variable columns that need to be discretized.Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns.}

\item{discnum}{(optional) A vector of number of classes for discretization. Default is \code{2:15}.}

\item{discmethod}{(optional) A vector of methods for discretization,default is used
\code{c("sd","equal","pretty","quantile","fisher","headtails","maximum","box")}in \code{spEcula}.}

\item{cores}{(optional) A positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}

\item{...}{(optional) Other arguments passed to \code{gd_bestunidisc()}.}
}
\value{
A nested tibble with \code{spatial_units}, \code{sesu_result} and \code{data}.
}
\description{
Function for comparison of size effects of spatial units in spatial heterogeneity analysis based on
optimal parameters geographical detector(OPGD) model.
}
\examples{
\dontrun{
library(tidyverse)
fvcpath = "https://github.com/SpatLyu/rdevdata/raw/main/FVC.tif"
fvc = terra::rast(paste0("/vsicurl/",fvcpath))
fvc1000 = fvc \%>\%
  terra::as.data.frame(na.rm = T) \%>\%
  as_tibble()
fvc5000 = fvc \%>\%
  terra::aggregate(fact = 5) \%>\%
  terra::as.data.frame(na.rm = T) \%>\%
  as_tibble()
gd_sesu(fvc ~ .,
        datalist = list(fvc1000,fvc5000),
        su = c(1000,5000),
        discnum = 2:15,
        discvar = names(select(fvc5000,-c(fvc,lulc))),
        cores = 6)
}
}
\references{
Song, Y., Wang, J., Ge, Y. & Xu, C. (2020) An optimal parameters-based geographical detector
model enhances geographic characteristics of explanatory variables for spatial heterogeneity
analysis: Cases with different types of spatial data, GIScience & Remote Sensing, 57(5), 593-610.
doi: 10.1080/15481603.2020.1760434.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
