% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_permutations}
\alias{gen_permutations}
\title{generate permutations}
\usage{
gen_permutations(x, seed = 123456789)
}
\arguments{
\item{x}{A vector.}

\item{seed}{(optional) Random seed number. Default is \code{123456789}.}
}
\value{
A permutations vector.
}
\description{
generate permutations
}
\examples{
gen_permutations(1:100,42)

}
