% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_py.R
\name{install_rpt}
\alias{install_rpt}
\title{set up \code{gdverse} python dependence}
\usage{
install_rpt(py_env = "r-gdverse", py_version = 3.9, method = "auto", ...)
}
\arguments{
\item{py_env}{The name, or full path, of the environment in which Python packages are to be installed.}

\item{py_version}{(optional) The requested Python version.Default is 3.9.}

\item{method}{(optional) Installation method. By default, "auto" automatically finds a method that will
work in the local environment.}

\item{...}{Other parameers passed to \code{reticulate::py_install()}}
}
\description{
Create an isolated Python virtual environment with all \code{gdverse} dependencies.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
