% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{loess_optscale}
\alias{loess_optscale}
\title{determine optimal spatial data analysis scale}
\usage{
loess_optscale(qvec, spscalevec, increase_rate = 0.05)
}
\arguments{
\item{qvec}{A numeric vector of q statistics.}

\item{spscalevec}{A numeric vector of spatial scales corresponding to \code{qvec}.}

\item{increase_rate}{(optional) The critical increase rate of the number of discretization.
Default is \verb{5\%}.}
}
\value{
A optimal number of spatial scale
}
\description{
Function for determining optimal spatial data analysis scale based on locally
estimated scatterplot smoothing (LOESS) model.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
