% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndvi}
\alias{ndvi}
\title{dataset of NDVI changes and its influencing factors}
\format{
\code{ndvi}: A tibble with 713 rows and 7 variables
}
\usage{
ndvi
}
\description{
dataset of NDVI changes and its influencing factors, modified from \code{GD} package.
}
\keyword{dataset}
\keyword{ndvi}
