% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NTDs}
\alias{NTDs}
\title{NTDs data}
\format{
\code{NTDs}: A tibble with 185 rows and 4 variable columns and 2 location columns, modified from \code{geodetector} package.
}
\usage{
NTDs
}
\description{
The data were obtained by preprocessing use \code{sf} and \code{tidyverse}.
}
\keyword{NTDs}
\keyword{dataset}
