% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsa.R
\name{pid_idsa}
\alias{pid_idsa}
\title{IDSA Q-saistics}
\usage{
pid_idsa(formula, rawdata, discdata, wt, overlaymethod = "and")
}
\arguments{
\item{formula}{A formula for IDSA Q-saistics \code{PID}}

\item{rawdata}{Raw observation data}

\item{discdata}{Discrete explanatory variables data}

\item{wt}{Spatial weight matrix}

\item{overlaymethod}{(optional) Spatial overlay method. One of \code{and}, \code{or}, \code{intersection}.
Default is \code{and}.}
}
\value{
The value of IDSA Q-saistics \code{PID}.
}
\description{
IDSA Q-saistics
}
\details{
\eqn{Q_{IDSA} = \frac{\theta_r}{\phi}}
}
\examples{
\dontrun{
library(sf)
usfi = read_sf(system.file('extdata/USFI_Xian.gpkg',package = 'gdverse')) |>
  dplyr::select(dplyr::all_of(c("NDVI","BH","WAR","SUHI")))
coord = usfi |>
  st_centroid() |>
  st_coordinates()
wt = inverse_distance_weight(coord[,1],coord[,2])
usf = usfi |>
  st_drop_geometry() |>
  dplyr::mutate(dplyr::across(1:3,\(.x) st_unidisc(.x,12)))
pid_idsa('NDVI~.',rawdata = usfi,discdata = usf,wt = wt)
}
}
