% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.R
\name{plot.ecological_detector}
\alias{plot.ecological_detector}
\title{plot ecological detector}
\usage{
\method{plot}{ecological_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{geodetector()}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for ecological detector in \code{geodetector()}.
}
