% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.factor_detector}
\alias{plot.factor_detector}
\title{plot factor detector result}
\usage{
\method{plot}{factor_detector}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{Return by \code{gd()}.}

\item{alpha}{(optional) Confidence level of the interval,default is \code{0.95}.}

\item{...}{Other arguments passed to \code{ggpubr::ggdotchart()}.}
}
\value{
A ggplot2 layer using \code{ggpubr::ggdotchart()}
}
\description{
S3 method to plot output for factor detector in \code{gd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
