% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gozh.R
\name{plot.gozh_result}
\alias{plot.gozh_result}
\title{plot GOZH result}
\usage{
\method{plot}{gozh_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{gozh()}.}

\item{...}{(optional) Other arguments.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for GOZH model result in \code{gozh()}.
}
