% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsa.R
\name{plot.idsa_result}
\alias{plot.idsa_result}
\title{plot IDSA risk result}
\usage{
\method{plot}{idsa_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{idsa()}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for IDSA risk result in \code{idsa()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
