% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.R
\name{plot.interaction_detector}
\alias{plot.interaction_detector}
\title{plot interaction detector result}
\usage{
\method{plot}{interaction_detector}(x, alpha = 1, ...)
}
\arguments{
\item{x}{Return by \code{geodetector()}.}

\item{alpha}{(optional) Picture transparency. Default is \code{1}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for interaction detector in \code{geodetector()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
