% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isp.R
\name{plot.isp_result}
\alias{plot.isp_result}
\title{plot ISP result}
\usage{
\method{plot}{isp_result}(x, low_color = "#6600CC", high_color = "#FFCC33", ...)
}
\arguments{
\item{x}{Return by \code{isp()}.}

\item{low_color}{(optional) The low color of the color gradient, default is \verb{#6600CC}.}

\item{high_color}{(optional) The high color of the color gradient, default is \verb{#FFCC33}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for ISP result in \code{isp()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
