% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesh.R
\name{plot.lesh_result}
\alias{plot.lesh_result}
\title{plot LESH model result}
\usage{
\method{plot}{lesh_result}(x, pie = TRUE, scatter = FALSE, scatter_alpha = 1, n_pielegend = 3, ...)
}
\arguments{
\item{x}{x Return by \code{lesh()}.}

\item{pie}{(optional) Whether to draw the interaction contributions. Default is \code{TRUE}.}

\item{scatter}{(optional) Whether to draw the interaction direction diagram. Default is \code{FALSE}.}

\item{scatter_alpha}{(optional) Picture transparency. Default is \code{1}.}

\item{n_pielegend}{(optional) The number of interaction contributions pie plot legend. Default is \code{3}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer.
}
\description{
S3 method to plot output for LESH model interaction result in \code{lesh()}.
}
\note{
When both \code{scatter} and \code{pie} are set to \code{TRUE} in RStudio, enlarge the drawing frame
for normal display.
}
\examples{
\dontrun{
data('ndvi')
g = lesh(NDVIchange ~ ., data = ndvi, cores = 6)
plot(g)
}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
