% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opgd.R
\name{plot.opgd_result}
\alias{plot.opgd_result}
\title{plot OPGD result}
\usage{
\method{plot}{opgd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{opgd()}.}

\item{...}{(optional) Other arguments.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for OPGD model result in \code{opgd()}.
}
