% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgd.R
\name{plot.rgd_result}
\alias{plot.rgd_result}
\title{plot RGD result}
\usage{
\method{plot}{rgd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{rgd()}.}

\item{...}{(optional) Other arguments passed to \code{patchwork::wrap_plots()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for RGD model result in \code{rgd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
