% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rid.R
\name{plot.rid_result}
\alias{plot.rid_result}
\title{plot RID result}
\usage{
\method{plot}{rid_result}(x, alpha = 1, ...)
}
\arguments{
\item{x}{Return by \code{rid()}.}

\item{alpha}{(optional) Picture transparency. Default is \code{1}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for RID model from \code{rid()}.
}
