% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesu_gozh.R
\name{plot.sesu_gozh}
\alias{plot.sesu_gozh}
\title{plot gozh sesu}
\usage{
\method{plot}{sesu_gozh}(x, ...)
}
\arguments{
\item{x}{Return by \code{sesu_gozh()}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer.
}
\description{
S3 method to plot output for gozh sesu in \code{sesu_gozh()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
