% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesu_opgd.R
\name{plot.sesu_opgd}
\alias{plot.sesu_opgd}
\title{plot opgd sesu}
\usage{
\method{plot}{sesu_opgd}(x, ...)
}
\arguments{
\item{x}{Return by \code{sesu_opgd()}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer.
}
\description{
S3 method to plot output for opgd sesu in \code{sesu_opgd()}.
}
