% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade.R
\name{plot.spade_result}
\alias{plot.spade_result}
\title{plot SPADE power of spatial and multilevel discretization determinant}
\usage{
\method{plot}{spade_result}(x, slicenum = 2, alpha = 0.95, ...)
}
\arguments{
\item{x}{Return by \code{spade()}.The number of labels facing inward.}

\item{slicenum}{(optional) The number of labels facing inward. Default is \code{2}.}

\item{alpha}{(optional) Confidence level.Default is \code{0.95}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer.
}
\description{
S3 method to plot output for SPADE power of spatial and multilevel discretization
determinant from \code{spade()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
