% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_geodetecor.R
\name{plot.srs_ecological_detector}
\alias{plot.srs_ecological_detector}
\title{plot spatial rough set-based ecological detector}
\usage{
\method{plot}{srs_ecological_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{srsgd()}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for spatial rough set-based ecological detector in \code{srsgd()}.
}
