% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_geodetecor.R
\name{plot.srs_interaction_detector}
\alias{plot.srs_interaction_detector}
\title{plot spatial rough set-based interaction detector result}
\usage{
\method{plot}{srs_interaction_detector}(x, alpha = 1, ...)
}
\arguments{
\item{x}{Return by \code{srsgd()}.}

\item{alpha}{(optional) Picture transparency. Default is \code{1}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for spatial rough set-based interaction detector in \code{srsgd()}.
}
