% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srsgd.R
\name{plot.srsgd_result}
\alias{plot.srsgd_result}
\title{plot SRSGD result}
\usage{
\method{plot}{srsgd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{srsgd()}.}

\item{...}{(optional) Other arguments.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for SRSGD model result in \code{srsgd()}.
}
