% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp.R
\name{print.esp_result}
\alias{print.esp_result}
\title{print ESP result}
\usage{
\method{print}{esp_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{esp()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for ESP model from \code{esp()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
