% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{print.factor_detector}
\alias{print.factor_detector}
\title{print factor detector}
\usage{
\method{print}{factor_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{gd()}.}

\item{...}{Other arguments.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for factor detector in \code{gd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
