% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{print.gd_result}
\alias{print.gd_result}
\title{print GD result}
\usage{
\method{print}{gd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{gd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for GD model from \code{gd()}.
}
