% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gozh.R
\name{print.gozh_result}
\alias{print.gozh_result}
\title{print GOZH result}
\usage{
\method{print}{gozh_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{gozh()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for GOZH model from \code{gozh()}.
}
