% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsa.R
\name{print.idsa_result}
\alias{print.idsa_result}
\title{print IDSA result}
\usage{
\method{print}{idsa_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{idsa()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for IDSA model from \code{idsa()}.
}
