% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{print.interaction_detector}
\alias{print.interaction_detector}
\title{print interaction detector}
\usage{
\method{print}{interaction_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{gd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for interaction detector in \code{gd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
