% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesh.R
\name{print.interaction_result_lesh}
\alias{print.interaction_result_lesh}
\title{print LESH model interaction result}
\usage{
\method{print}{interaction_result_lesh}(x, ...)
}
\arguments{
\item{x}{Return by \code{lesh()}.}

\item{...}{Other arguments.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for interaction_result in \code{lesh()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
