% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isp.R
\name{print.isp_result}
\alias{print.isp_result}
\title{print ISP result}
\usage{
\method{print}{isp_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{isp()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for ISP model from \code{isp()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
