% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesh.R
\name{print.lesh_result}
\alias{print.lesh_result}
\title{print LESH model interaction result}
\usage{
\method{print}{lesh_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{lesh()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for LESH model interaction result in \code{lesh()}.
}
