% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opgd.R
\name{print.opgd_result}
\alias{print.opgd_result}
\title{print OPGD result}
\usage{
\method{print}{opgd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{opgd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for RID model from \code{opgd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
