% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resh.R
\name{print.resh_result}
\alias{print.resh_result}
\title{print RESH result}
\usage{
\method{print}{resh_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{resh()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for RESH model from \code{resh()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
