% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgd.R
\name{print.rgd_result}
\alias{print.rgd_result}
\title{print RGD result}
\usage{
\method{print}{rgd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{rgd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for RGD model from \code{rgd()}.
}
