% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rid.R
\name{print.rid_result}
\alias{print.rid_result}
\title{print RID result}
\usage{
\method{print}{rid_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{rid()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for RID model from \code{rid()}.
}
