% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.R
\name{print.risk_detector}
\alias{print.risk_detector}
\title{print risk detector}
\usage{
\method{print}{risk_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{geodetector()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for risk detector in \code{geodetector()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
