% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsh.R
\name{print.rsh_result}
\alias{print.rsh_result}
\title{print RSH result}
\usage{
\method{print}{rsh_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{rsh()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for RSH model from \code{rsh()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
