% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesu_gozh.R
\name{print.sesu_gozh}
\alias{print.sesu_gozh}
\title{print gozh sesu}
\usage{
\method{print}{sesu_gozh}(x, ...)
}
\arguments{
\item{x}{Return by \code{sesu_gozh()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for gozh sesu from \code{sesu_gozh()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
