% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesu_opgd.R
\name{print.sesu_opgd}
\alias{print.sesu_opgd}
\title{print opgd sedu}
\usage{
\method{print}{sesu_opgd}(x, ...)
}
\arguments{
\item{x}{Return by \code{sesu_opgd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for iopgd sedu in \code{sesu_opgd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
