% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade.R
\name{print.spade_result}
\alias{print.spade_result}
\title{print SPADE power of spatial and multilevel discretization determinant}
\usage{
\method{print}{spade_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{spade()}.}

\item{...}{Other arguments.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for SPADE power of spatial and multilevel discretization
determinant from \code{spade()}.
}
