% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_geodetecor.R
\name{print.srs_factor_detector}
\alias{print.srs_factor_detector}
\title{print spatial rough set-based factor detector}
\usage{
\method{print}{srs_factor_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{srsgd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for spatial rough set-based factor detector in \code{srsgd()}.
}
