% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srsgd.R
\name{print.srs_interaction_detector}
\alias{print.srs_interaction_detector}
\title{print spatial rough set-based interaction detector}
\usage{
\method{print}{srs_interaction_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{srsgd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for spatial rough set-based interaction detector in \code{srsgd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
