% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srsgd.R
\name{print.srsgd_result}
\alias{print.srsgd_result}
\title{print SRSGD result}
\usage{
\method{print}{srsgd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{srsgd()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for SRSGD model from \code{srsgd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
