% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{rescale_vector}
\alias{rescale_vector}
\title{rescale continuous vector to specified minimum and maximum}
\usage{
rescale_vector(x, to_left = 0, to_right = 1)
}
\arguments{
\item{x}{A continuous numeric vector.}

\item{to_left}{(optional) Specified minimum. Default is \code{0}.}

\item{to_right}{(optional) Specified maximum. Default is \code{1}.}
}
\value{
A continuous vector which has rescaled.
}
\description{
rescale continuous vector to specified minimum and maximum
}
\examples{
rescale_vector(c(-5,1,5),0.01,0.99)

}
