% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spEcula_helper.R
\name{shuffle_vector}
\alias{shuffle_vector}
\title{randomly shuffling vector}
\usage{
shuffle_vector(x, shuffle_rate, seed = 123456789)
}
\arguments{
\item{x}{A vector.}

\item{shuffle_rate}{The shuffling rate.}

\item{seed}{(optional) Random seed number. Default is \code{123456789}.}
}
\value{
A shuffled vector.
}
\description{
randomly shuffling vector
}
