% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulation data.}
\format{
\code{sim}: A tibble with 80 rows and 6 variables, modified from \code{IDSA} package.
}
\usage{
sim
}
\description{
Simulation data.
}
\keyword{dataset}
\keyword{sim}
