% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd_lesh.R
\name{spd_lesh}
\alias{spd_lesh}
\title{SHAP power of determinants (SPD)}
\usage{
spd_lesh(formula, data, cores = 1, ...)
}
\arguments{
\item{formula}{A formula of calculate SHAP power of determinants \eqn{SPD}.}

\item{data}{A data.frame or tibble of observation data.}

\item{cores}{(optional) A positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster object.}

\item{...}{(optional) Other arguments passed to \code{gozh()}.}
}
\value{
A tibble with variable and its corresponding \eqn{SPD} value.
}
\description{
SHAP power of determinants (SPD)
}
\examples{
\dontrun{
data('ndvi')
g = spd_lesh(NDVIchange ~ ., data = ndvi, cores = 6)
g
}
}
