% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srs_table}
\alias{srs_table}
\title{example of spatial information system table}
\format{
\code{srs_table}: A tibble with 11 rows and 5 variables(one \code{ID} column).
}
\usage{
srs_table
}
\description{
example of spatial information system table
}
\keyword{dataset}
\keyword{srs_table}
