% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srs_wt}
\alias{srs_wt}
\title{example of spatial information system spatial adjacency matrix}
\format{
\code{srs_wt}: A matrix with 11rows and 11cols.
}
\usage{
srs_wt
}
\description{
example of spatial information system spatial adjacency matrix
}
\keyword{dataset}
\keyword{srs_wt}
