% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyoverlay.R
\name{st_fuzzyoverlay}
\alias{st_fuzzyoverlay}
\title{spatial fuzzy overlay}
\usage{
st_fuzzyoverlay(formula, data, method = "and")
}
\arguments{
\item{formula}{A formula of spatial fuzzy overlay.}

\item{data}{A data.frame or tibble of observation data.}

\item{method}{(optional) Overlay methods. When \code{method} is \code{and}, use \code{min} to do
fuzzy overlay;and when \code{method} is \code{or},use \code{max} to do fuzzy overlay. Default is \code{and}.}
}
\value{
A spatial fuzzy overlay vector.
}
\description{
Function for spatial fuzzy overlay.
}
\references{
Yongze Song & Peng Wu (2021) An interactive detector for spatial associations,
International Journal of Geographical Information Science, 35:8, 1676-1701,
DOI:10.1080/13658816.2021.1882680
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
