% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{unidisc_methods}
\alias{unidisc_methods}
\title{all discretization methods that can be used in \code{st_unidisc}}
\usage{
unidisc_methods()
}
\value{
A character vector
}
\description{
A comprehensive vector of all discretization methods that can be
employed within \code{st_unidisc()}.
}
\examples{
unidisc_methods()

}
