% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_assign}
\alias{weight_assign}
\title{assign values by weight}
\usage{
weight_assign(x, w, list = FALSE)
}
\arguments{
\item{x}{A numeric value}

\item{w}{A weight vector}

\item{list}{(optional) Return list or not. if \code{list} is \code{TRUE}, return a list,
otherwise return a vector. Default is \code{FALSE}.}
}
\value{
A numeric Vector.
}
\description{
assign values by weight
}
\examples{
weight_assign(0.875,1:3)

}
