% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.R
\name{factor_detector}
\alias{factor_detector}
\title{factor detector}
\usage{
factor_detector(y, x)
}
\arguments{
\item{y}{Variable Y, continuous numeric vector.}

\item{x}{Covariate X, \code{factor}, \code{character} or \code{discrete numeric}.}
}
\value{
A list contains the Q-statistic and the p-value.
}
\description{
The factor detector q-statistic measures the spatial stratified heterogeneity of a variable Y,
or the determinant power of a covariate X of Y.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
