% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{gd}
\alias{gd}
\title{geographical detector model}
\usage{
gd(formula, data, type = "factor", ...)
}
\arguments{
\item{formula}{A formula of geographical detector model.}

\item{data}{A data.frame or tibble of observation data.}

\item{type}{(optional) The type of geographical detector,which must be one of \code{factor}(default),
\code{interaction}, \code{risk}, \code{ecological}.}

\item{...}{(optional) Specifies the size of the alpha (confidence level).Default is \code{0.95}.}
}
\value{
A tibble of the corresponding result is stored under the corresponding detector type.
}
\description{
Classical geographical detector model.
}
\examples{
gd(y ~ x1 + x2,
   tibble::tibble(y = 1:7,
                  x1 = c('x',rep('y',3),rep('z',3)),
                  x2 = c(rep('a',2),rep('b',2),rep('c',3))))

gd(y ~ x1 + x2,
   tibble::tibble(y = 1:7,
                  x1 = c('x',rep('y',3),rep('z',3)),
                  x2 = c(rep('a',2),rep('b',2),rep('c',3))),
   type = 'interaction')

gd(y ~ x1 + x2,
   tibble::tibble(y = 1:7,
                  x1 = c('x',rep('y',3),rep('z',3)),
                  x2 = c(rep('a',2),rep('b',2),rep('c',3))),
   type = 'risk',alpha = 0.95)

gd(y ~ x1 + x2,
   tibble::tibble(y = 1:7,
                  x1 = c('x',rep('y',3),rep('z',3)),
                  x2 = c(rep('a',2),rep('b',2),rep('c',3))),
   type = 'ecological',alpha = 0.95)

}
\references{
Jin‐Feng Wang, Xin‐Hu Li, George Christakos, Yi‐Lan Liao, Tin Zhang, XueGu & Xiao‐Ying Zheng (2010) Geo
graphical Detectors‐Based Health Risk Assessment and its Application in the Neural Tube Defects Study of the
Heshun Region, China, International Journal of Geographical Information Science, 24:1, 107-127, DOI: 10.1080/13658810802443457
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
