% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_bestunidisc.R
\name{gd_bestunidisc}
\alias{gd_bestunidisc}
\title{best univariate discretization based on geodetector q-statistic}
\usage{
gd_bestunidisc(
  formula,
  data,
  discnum = 3:8,
  discmethod = c("sd", "equal", "geometric", "quantile", "natural"),
  cores = 1,
  return_disc = TRUE,
  seed = 123456789,
  ...
)
}
\arguments{
\item{formula}{A formula of best univariate discretization.}

\item{data}{A \code{data.frame} or \code{tibble} of observation data.}

\item{discnum}{(optional) A vector of number of classes for discretization. Default is \code{3:8}.}

\item{discmethod}{(optional) A vector of methods for discretization, default is using
\code{c("sd","equal","geometric","quantile","natural")} by invoking \code{sdsfun}.}

\item{cores}{(optional) Positive integer (default is 1). When cores are greater than 1, use
multi-core parallel computing.}

\item{return_disc}{(optional) Whether or not return discretized result used the optimal parameter.
Default is \code{TRUE}.}

\item{seed}{(optional) Random seed number, default is \code{123456789}.}

\item{...}{(optional) Other arguments passed to \code{sdsfun::discretize_vector()}.}
}
\value{
A list.
\describe{
\item{\code{x}}{the name of the variable that needs to be discretized}
\item{\code{k}}{optimal discretization number}
\item{\code{method}}{optimal discretization method}
\item{\code{disc}}{optimal discretization results}
}
}
\description{
Function for determining the best univariate discretization based on geodetector q-statistic.
}
\examples{
data('sim')
gd_bestunidisc(y ~ xa + xb + xc,
               data = sim,
               discnum = 3:6)

}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
