% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gozh.R
\name{gozh}
\alias{gozh}
\title{geographically optimal zones-based heterogeneity(GOZH) model}
\usage{
gozh(formula, data, cores = 1, ...)
}
\arguments{
\item{formula}{A formula of GOZH model.}

\item{data}{A data.frame or tibble of observation data.}

\item{cores}{positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}

\item{...}{(optional) Other arguments passed to \code{rpart::rpart}.}
}
\value{
A list of the GOZH model result.
}
\description{
Function for geographically optimal zones-based heterogeneity(GOZH) model
}
\examples{
\dontrun{
ndvi = GD::ndvi_40
g = gozh(NDVIchange ~ ., data = ndvi)
g
}
}
\references{
Luo, P., Song, Y., Huang, X., Ma, H., Liu, J., Yao, Y., & Meng, L. (2022). Identifying determinants of
spatio-temporal disparities in soil moisture of the Northern Hemisphere using a geographically optimal
zones-based heterogeneity model. ISPRS Journal of Photogrammetry and Remote Sensing: Official
Publication of the International Society for Photogrammetry and Remote Sensing (ISPRS), 185, 111–128.
https://doi.org/10.1016/j.isprsjprs.2022.01.009
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
