% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idsa.R
\name{idsa}
\alias{idsa}
\title{interactive detector for spatial associations(IDSA)}
\usage{
idsa(
  formula,
  data,
  wt = NULL,
  overlaymethod = "and",
  locations = NULL,
  discvar = NULL,
  discnum = NULL,
  discmethod = NULL,
  strategy = 2L,
  increase_rate = 0.05,
  cores = 6,
  seed = 123456789,
  ...
)
}
\arguments{
\item{formula}{A formula of IDSA model.}

\item{data}{A data.frame or tibble of observation data.}

\item{wt}{(optional) The spatial weight matrix.When \code{wt} is not provided, must provide \code{locations}.
And \code{gdverse} will use \code{locations} columns to construct spatial weight use \code{inverse_distance_weight()}.}

\item{overlaymethod}{(optional) Spatial overlay method. One of \code{and}, \code{or}, \code{intersection}.
Default is \code{and}.}

\item{locations}{(optional) The geospatial locations coordinate columns name which in \code{data}.
Useful and must provided when \code{wt} is not provided. When \code{wt} is provided, \code{locations} is not need.}

\item{discvar}{Name of continuous variable columns that need to be discretized.Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns.}

\item{discnum}{(optional) Number of multilevel discretization.Default will use \code{3:22}.}

\item{discmethod}{(optional) The discretization methods. Default all use \code{quantile}. More details to see \code{st_unidisc()}.}

\item{strategy}{(optional) Discretization strategy. When \code{strategy} is \code{1L}, choose the highest SPADE model q-statistics to
determinate optimal spatial data discretization parameters. When \code{strategy} is \code{2L}, The optimal discrete parameters of
spatial data are selected by combining LOESS model.}

\item{increase_rate}{(optional) The critical increase rate of the number of discretization.
Default is \verb{5\%}.}

\item{cores}{(optional) A positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}

\item{seed}{(optional) Random number seed, default is \code{123456789}.}

\item{...}{(optional) Other arguments passed to \code{cpsd_disc()}.}
}
\value{
A tibble with PID values under different spatial overlays.
}
\description{
Function for interactive detector for spatial associations model.
}
\examples{
\dontrun{
library(sf)
usfi = read_sf(system.file('extdata/USFI_Xian.gpkg',package = 'gdverse')) |>
  dplyr::select(dplyr::all_of(c("NDVI","BH","SUHI")))
coord = usfi |>
  st_centroid() |>
  st_coordinates()
wt = inverse_distance_weight(coord[,1],coord[,2])
usfi = st_drop_geometry(usfi)
idsa(SUHI ~ NDVI + BH, data = usfi,wt = wt,
     discvar = c('NDVI','BH'),cores = 6)
}
}
\references{
Yongze Song & Peng Wu (2021) An interactive detector for spatial associations,
International Journal of Geographical Information Science, 35:8, 1676-1701,
DOI:10.1080/13658816.2021.1882680
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
