% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.R
\name{interaction_detector}
\alias{interaction_detector}
\title{interaction detector}
\usage{
interaction_detector(y, x1, x2)
}
\arguments{
\item{y}{Dependent variable, continuous numeric vector.}

\item{x1}{Covariable \eqn{X_1}, \code{factor}, \code{character} or \code{discrete numeric}.}

\item{x2}{Covariable \eqn{X_2}, \code{factor}, \code{character} or \code{discrete numeric}.}
}
\value{
A list contains the Q statistic when the factors \eqn{X_1} and \eqn{X_1} act on \eqn{Y} alone
and the Q statistic when the two interact on \eqn{Y} together with the result type of the interaction detector.
}
\description{
Identify the interaction between different risk factors, that is, assess whether factors X1 and X2 together
increase or decrease the explanatory power of the dependent variable Y, or whether the effects of these factors
on Y are independent of each other.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
