% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opgd.R
\name{opgd}
\alias{opgd}
\title{optimal parameters-based geographical detector(OPGD) model}
\usage{
opgd(
  formula,
  data,
  discvar,
  discnum = NULL,
  discmethod = NULL,
  cores = 1,
  type = "factor",
  alpha = 0.95,
  ...
)
}
\arguments{
\item{formula}{A formula of OPGD model.}

\item{data}{A data.frame or tibble of observation data.}

\item{discvar}{Name of continuous variable columns that need to be discretized.Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns.}

\item{discnum}{(optional) A vector of number of classes for discretization. Default is \code{3:15}.}

\item{discmethod}{(optional) A vector of methods for discretization,default is using
\code{c("sd","equal","pretty","quantile","fisher","headtails","maximum","box")}in \code{spEcula}.}

\item{cores}{(optional) A positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}

\item{type}{(optional) The type of geographical detector,which must be \code{factor}(default),
\code{interaction}, \code{risk}, \code{ecological}.You can run one or more at a time.}

\item{alpha}{(optional) Specifies the size of confidence level.Default is \code{0.95}.}

\item{...}{(optional) Other arguments passed to \code{gd_bestunidisc()}.A useful parameter is \code{seed},
which is used to set the random number seed.}
}
\value{
A list of the OPGD model result.
}
\description{
Function for optimal parameters-based geographical detector(OPGD) model.
}
\examples{
\dontrun{
library(terra)
library(tidyverse)
fvcpath = "https://github.com/SpatLyu/rdevdata/raw/main/FVC.tif"
fvc = terra::rast(paste0("/vsicurl/",fvcpath))
fvc = terra::aggregate(fvc,fact = 5)
fvc = as_tibble(terra::as.data.frame(fvc,na.rm = T))
opgd(fvc ~ ., data = fvc,
     discvar = names(select(fvc,-c(fvc,lulc))),
     cores = 6, type =c('factor','interaction'))
}
}
\references{
Song, Y., Wang, J., Ge, Y. & Xu, C. (2020) An optimal parameters-based geographical detector
model enhances geographic characteristics of explanatory variables for spatial heterogeneity
analysis: Cases with different types of spatial data, GIScience & Remote Sensing, 57(5), 593-610.
doi: 10.1080/15481603.2020.1760434.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
