% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp.R
\name{plot.esp_result}
\alias{plot.esp_result}
\title{plot ESP result}
\usage{
\method{plot}{esp_result}(x, low_color = "#6600CC", high_color = "#FFCC33", ...)
}
\arguments{
\item{x}{Return by \code{esp()}.}

\item{low_color}{(optional) The low color of the color gradient, defaul is \verb{#6600CC}.}

\item{high_color}{(optional) The high color of the color gradient, defaul is \verb{#FFCC33}.}

\item{...}{(optional) Other arguments passed to \code{ggplot2::theme()}.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for ESP result in \code{esp()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
