% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{plot.gd_result}
\alias{plot.gd_result}
\title{plot GD result}
\usage{
\method{plot}{gd_result}(x, ...)
}
\arguments{
\item{x}{Return by \code{gd()}.}

\item{...}{(optional) Other arguments.}
}
\value{
A ggplot2 layer
}
\description{
S3 method to plot output for GD model result in \code{gd()}.
}
