% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.R
\name{print.ecological_detector}
\alias{print.ecological_detector}
\title{print ecological detector}
\usage{
\method{print}{ecological_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{geodetector()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for ecological detector in \code{geodetector()}.
}
