% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodetector.R
\name{print.factor_detector}
\alias{print.factor_detector}
\title{print factor detector}
\usage{
\method{print}{factor_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{geodetector()}.}

\item{...}{(optional) Other arguments passed to \code{knitr::kable()}.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for factor detector in \code{geodetector()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
