% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd.R
\name{print.risk_detector}
\alias{print.risk_detector}
\title{print risk detector}
\usage{
\method{print}{risk_detector}(x, ...)
}
\arguments{
\item{x}{Return by \code{gd()}.}

\item{...}{Other arguments.}
}
\value{
Formatted string output
}
\description{
S3 method to format output for risk detector in \code{gd()}.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
