% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudop.R
\name{psmd_pseudop}
\alias{psmd_pseudop}
\title{power of spatial and multilevel discretization determinant(PSMD) and the corresponding pseudo-p value}
\usage{
psmd_pseudop(
  formula,
  data,
  wt = NULL,
  locations = NULL,
  discnum = NULL,
  discmethod = NULL,
  cores = 6,
  seed = 123456789,
  permutations = 99,
  ...
)
}
\arguments{
\item{formula}{A formula of calculate power of spatial and multilevel discretization determinant \eqn{PSMDQ_s}.}

\item{data}{A data.frame or tibble of observation data.}

\item{wt}{(optional) The spatial weight matrix.When \code{wt} is not provided, must provide \code{locations}.
And \code{gdverse} will use \code{locations} columns to construct spatial weight use \code{inverse_distance_weight()}.}

\item{locations}{(optional) The geospatial locations coordinate columns name which in \code{data}.
Useful and must provided when \code{wt} is not provided.}

\item{discnum}{(optional) Number of multilevel discretization.Default will use \code{3:22}.}

\item{discmethod}{(optional) The discretization methods. Default will use \code{quantile}.
When \code{discmethod} is \code{robust} use \code{robust_disc()}, others use \code{st_unidisc()}.Now only support
one \code{discmethod} at one time.}

\item{cores}{(optional) A positive integer(default is 6). If cores > 1, use parallel computation.}

\item{seed}{(optional) Random seed number, default is \code{123456789}.}

\item{permutations}{(optional) The number of permutations for the PSD computation. Default is \code{99}.
If \code{permutations} is \code{0}, no pseudo-p values are calculated.}

\item{...}{(optional) Other arguments passed to \code{st_unidisc()} or \code{robust_disc()}.}
}
\value{
A tibble of power of spatial and multilevel discretization determinant and the corresponding pseudo-p value.
}
\description{
Function for calculate power of spatial and multilevel discretization determinant and the corresponding pseudo-p value.
}
\details{
The power of spatial and multilevel discretization determinant formula is
\eqn{PSMDQ_s = MEAN(Q_s)}
}
\examples{
\dontrun{
library(sf)
usfi = read_sf(system.file('extdata/USFI_Xian.gpkg',package = 'gdverse')) |>
  dplyr::select(dplyr::all_of(c("NDVI","BH","SUHI")))
coord = usfi |>
  st_centroid() |>
  st_coordinates()
usfi = usfi |>
  dplyr::bind_cols(coord) |>
  st_drop_geometry()
tictoc::tic()
pp = psmd_pseudop('SUHI ~ BH',data = dplyr::select(usfi,SUHI,BH,X,Y),
                  locations = c('X','Y'),cores = 6)
tictoc::toc()
pp
}
}
\references{
Xuezhi Cang & Wei Luo (2018) Spatial association detector (SPADE),International
Journal of Geographical Information Science, 32:10, 2055-2075, DOI:  10.1080/13658816.2018.1476693
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
