% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_spade.R
\name{psmd_spade}
\alias{psmd_spade}
\title{power of spatial and multilevel discretization determinant(PSMD)}
\usage{
psmd_spade(
  formula,
  data,
  wt = NULL,
  locations = NULL,
  discnum = NULL,
  discmethod = "quantile",
  cores = 1,
  seed = 123456789,
  ...
)
}
\arguments{
\item{formula}{A formula of spatial stratified heterogeneity test.}

\item{data}{A data.frame or tibble of observation data.}

\item{wt}{(optional) The spatial weight matrix.When \code{wt} is not provided, must provide \code{locations}.
And gdverse will use \code{locations} columns to construct spatial weight use \code{inverse_distance_weight()}.}

\item{locations}{(optional) The geospatial locations coordinate columns name which in \code{data}.
Useful and must provided when \code{wt} is not provided.}

\item{discnum}{(optional) Number of multilevel discretization.Default will use \code{3:15}.}

\item{discmethod}{(optional) The discretization methods. Default will use \code{quantile}.
When \code{discmethod} is \code{robust} use \code{robust_disc()}, others use \code{st_unidisc()}.}

\item{cores}{positive integer(default is 1). If cores > 1, a 'parallel' package
cluster with that many cores is created and used. You can also supply a cluster
object.}

\item{seed}{(optional) Random seed number, default is \code{123456789}.}

\item{...}{(optional) Other arguments passed to \code{st_unidisc()} or \code{robust_disc()}.}
}
\value{
A value of power of spatial and multilevel discretization determinant \eqn{PSMDQ_s}.
}
\description{
Function for calculate power of spatial and multilevel discretization determinant \eqn{PSMDQ_s}.
}
\details{
The power of spatial and multilevel discretization determinant formula is
\eqn{PSMDQ_s = MEAN(Q_s)}
}
\references{
Xuezhi Cang & Wei Luo (2018) Spatial association detector (SPADE),International
Journal of Geographical Information Science, 32:10, 2055-2075, DOI:  10.1080/13658816.2018.1476693
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
