% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgd.R
\name{rgd}
\alias{rgd}
\title{robust geographical detector(RGD) model}
\usage{
rgd(
  formula,
  data,
  discvar,
  discnum = NULL,
  minsize = NULL,
  cores = 1,
  type = "factor",
  alpha = 0.95,
  ...
)
}
\arguments{
\item{formula}{A formula of RGD model.}

\item{data}{A data.frame or tibble of observation data.}

\item{discvar}{Name of continuous variable columns that need to be discretized.Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns.}

\item{discnum}{A numeric vector of discretized classes of columns that need to be discretized.
Default all \code{discvar} use \code{10}.}

\item{minsize}{(optional) The min size of each discretization group.Default all use \code{1}.}

\item{cores}{positive integer(default is 1). If cores > 1, use \code{python}'s \code{joblib} package to
parallel computation.}

\item{type}{(optional) The type of geographical detector,which must be \code{factor}(default),
\code{interaction}, \code{risk}, \code{ecological}.You can run one or more at a time.}

\item{alpha}{(optional) Specifies the size of confidence level.Default is \code{0.95}.}

\item{...}{(optional) Other arguments passed to \code{robust_disc()}.}
}
\value{
A list of the RGD model result.
}
\description{
Function for robust geographical detector(RGD) model.
}
\examples{
\dontrun{
library(terra)
library(tidyverse)
fvcpath = "https://github.com/SpatLyu/rdevdata/raw/main/FVC.tif"
fvc = terra::rast(paste0("/vsicurl/",fvcpath))
fvc = terra::aggregate(fvc,fact = 5)
fvc = as_tibble(terra::as.data.frame(fvc,na.rm = T))
rgd(fvc ~ ., data = fvc, discnum = 10,
    discvar = names(select(fvc,-c(fvc,lulc))),
    cores = 6, type =c('factor','interaction'))
}
}
\references{
Zhang, Z., Song, Y.*, & Wu, P., 2022. Robust geographical detector. International Journal of Applied Earth Observation and Geoinformation. 109, 102782.
DOI: 10.1016/j.jag.2022.102782.
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
