% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rid.R
\name{rid}
\alias{rid}
\title{robust interaction detector(RID) model}
\usage{
rid(
  formula,
  data,
  discvar = NULL,
  discnum = 3:8,
  minsize = 1,
  strategy = 2L,
  increase_rate = 0.05,
  cores = 1
)
}
\arguments{
\item{formula}{A formula of RGD model.}

\item{data}{A \code{data.frame}, \code{tibble} or \code{sf} object of observation data.}

\item{discvar}{Name of continuous variable columns that need to be discretized. Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns. By default, all independent
variables are used as \code{discvar}.}

\item{discnum}{A numeric vector of discretized classes of columns that need to be discretized.
Default all \code{discvar} use \code{3:8}.}

\item{minsize}{(optional) The min size of each discretization group. Default all use \code{1}.}

\item{strategy}{(optional) Optimal discretization strategy. When \code{strategy} is \code{1L}, choose the highest
q-statistics to determinate optimal spatial data discretization parameters. When \code{strategy} is \code{2L},
The optimal discrete parameters of spatial data are selected by combining LOESS model.}

\item{increase_rate}{(optional) The critical increase rate of the number of discretization. Default is \verb{5\%}.}

\item{cores}{(optional) Positive integer (default is 1). When cores are greater than 1, use
multi-core parallel computing.}
}
\value{
A list.
\describe{
\item{\code{interaction}}{the result of RID model}
}
}
\description{
robust interaction detector(RID) model
}
\examples{
data('sim')
\donttest{
tryCatch({
  g = rid(y ~ .,
          data = dplyr::select(sim,-dplyr::any_of(c('lo','la'))),
          discnum = 3:6, cores = 1)
  g
}, error = \(e) message("Skipping Python-dependent example: ", e$message))
}
}
\references{
Zhang, Z., Song, Y., Karunaratne, L., & Wu, P. (2024). Robust interaction detector:
A case of road life expectancy analysis. Spatial Statistics, 59(100814), 100814.
https://doi.org/10.1016/j.spasta.2024.100814
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
