% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustdisc.R
\name{robust_disc}
\alias{robust_disc}
\title{univariate discretization based on offline change point detection.}
\usage{
robust_disc(formula, data, discnum, minsize = NULL, cores = 1)
}
\arguments{
\item{formula}{A formula of spatial stratified heterogeneity test.}

\item{data}{A data.frame or tibble of observation data.}

\item{discnum}{A numeric vector of discretized classes of columns that need to be discretized.}

\item{minsize}{(optional) The min size of each discretization group.Default all use \code{1}.}

\item{cores}{positive integer(default is 1). If cores > 1, use \code{python}'s \code{joblib} package to
parallel computation.}
}
\value{
A tibble of discretized classes of columns that need to be discretized.
}
\description{
Determines discretization interval breaks using an optimization algorithm for variance-based
change point detection.
}
\examples{
\dontrun{
library(terra)
library(tidyverse)
fvcpath = "https://github.com/SpatLyu/rdevdata/raw/main/FVC.tif"
fvc = terra::rast(paste0("/vsicurl/",fvcpath))
fvc = terra::aggregate(fvc,fact = 5)
fvc = as_tibble(terra::as.data.frame(fvc,na.rm = T))
new.fvc = robust_disc(fvc ~ .,data = select(fvc,-lulc),discnum = 10,cores = 6)
new.fvc
}
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
