% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade.R
\name{spade}
\alias{spade}
\title{spatial association detector (SPADE) model}
\usage{
spade(
  formula,
  data,
  wt = NULL,
  locations = NULL,
  discvar = NULL,
  discnum = NULL,
  discmethod = NULL,
  cores = 1,
  seed = 123456789,
  ...
)
}
\arguments{
\item{formula}{A formula of spatial association detector (SPADE) model.}

\item{data}{A data.frame or tibble of observation data.}

\item{wt}{(optional) The spatial weight matrix.When \code{wt} is not provided, must provide \code{locations}.
And \code{gdverse} will use \code{locations} columns to construct spatial weight use \code{inverse_distance_weight()}.}

\item{locations}{(optional) The geospatial locations coordinate columns name which in \code{data}.
Useful and must provided when \code{wt} is not provided.}

\item{discvar}{Name of continuous variable columns that need to be discretized.Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns.}

\item{discnum}{(optional) Number of multilevel discretization.Default will use \code{3:15}.}

\item{discmethod}{(optional) The discretization methods. Default all use \code{quantile}.
When \code{discmethod} is \code{robust} use \code{robust_disc()}, others use \code{st_unidisc()}}

\item{cores}{(optional) A positive integer(default is 1). If cores > 1, use parallel computation.}

\item{seed}{(optional) Random seed number, default is \code{123456789}.}

\item{...}{(optional) Other arguments passed to \code{st_unidisc()} or \code{robust_disc()}.}
}
\value{
A list of the SPADE model result.
}
\description{
Function for spatial association detector (SPADE) model.
}
\examples{
\dontrun{
library(sf)
usfi = read_sf(system.file('extdata/USFI_Xian.gpkg',package = 'gdverse')) |>
  dplyr::select(dplyr::all_of(c("NDVI","BH","SUHI")))
coord = usfi |>
  st_centroid() |>
  st_coordinates()
wt = inverse_distance_weight(coord[,1],coord[,2])
usfi = usfi |>
  dplyr::bind_cols(coord) |>
  st_drop_geometry()
spade('SUHI~.', data = usfi,locations = c('X','Y'),
      discvar = c('BH','NDVI'), cores = 6)
spade('SUHI~.', data = usfi, wt = wt,
      discvar = c('BH','NDVI'),
      discmethod = c('sd','equal'),cores = 6)
}
}
\references{
Xuezhi Cang & Wei Luo (2018) Spatial association detector (SPADE),International
Journal of Geographical Information Science, 32:10, 2055-2075, DOI:  10.1080/13658816.2018.1476693
}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
